add isPendingHash method
